Move: 공통 요소 타입 파일 생성 (common.ts)
